gwp_asan_headers is marked as available to platform
The header lib is depended on by many of the bionic libraries which are
part of the platform.

With b/153073816, we will be checking the availability to the platform,
i.e. platform-available module depending on platform-unavailable module
causes a build error.

Prepare for the change by correctly marking that this interface lib is
available to the platform.

Also, this change narrows down the APEXes where this header lib is
available for. Instead of allowing it to be included in any APEX, the
list of APEXes that are currently using this is specified.

Bug: 153073816
Test: m
Merged-In: Ie329792936599072aaf2bc66a51b64f064baa229
(cherry picked from commit 50332c5d6f14f080bcba05cc580776aa710a6851)
Change-Id: Ie329792936599072aaf2bc66a51b64f064baa229