[GWP-ASan] Fix flaky test on Fuchsia
The LateInit test might be reusing some already initialized thread
specific data if run within the main thread. This means that there
is a chance that the current value will not be enough for the 100
iterations, hence the test flaking.

Fix this by making the test run in its own thread.

Differential Revision: https://reviews.llvm.org/D92415

GitOrigin-RevId: c904c32b9c9243b11ffc18e46b7350f000e9c088
Change-Id: I9e7a0d56dd52d10a328296f509ab56aa5647ce9a